Feature/testing

Added:
- AppUserService Unit test (100% line coverage)
- AssignmentGroupService Unit test (100% line coverage)
- InteractionService Unit test (100% line coverage)
- AppUser endpoint Integration (4) Tests
- Authentication endpoint Integration (2) Tests
- Needed files for integration test: application-test & test-data.sql

Redid:
- Refactored AppUser phoneNumber from INT to Long because int is not nullable and in patch defaulting do 0.
- Exception handling for Spring Boot security, now throws appropriate exceptions via new CustomAuthenticationEntryPoint combined in SecurityConfig.

+++ src/main/java/nl/yourivb/TicketTrack/dtos/AppUser/AppUserDto.java
@@ -5,7 +5,7 @@ import java.time.LocalDateTime;
 public class AppUserDto {
     private Long id;
     private String name;
-    private int phoneNumber;
+    private Long phoneNumber;
     private String email;
     private String info;
     private LocalDateTime created;
@@ -29,11 +29,11 @@ public class AppUserDto {
         this.name = name;
     }
 
-    public int getPhoneNumber() {
+    public Long getPhoneNumber() {
         return phoneNumber;
     }
 
-    public void setPhoneNumber(int phoneNumber) {
+    public void setPhoneNumber(Long phoneNumber) {
         this.phoneNumber = phoneNumber;
     }
